Individual Behaviour (Fully Working) & Flocking (Nearly Working) Updates
Updated DNA attribute ranges for predator and prey to increase maxSpeed and maxForce. Refactored Flocking and IndividualBehaviour scenarios to use lists of behaviours for both predator and prey, replacing Evade and Pursuit with Flee and Seek/Wander combinations. Adjusted behaviour weights and logic for more dynamic and flexible agent responses.